# Conflicts:
#	src/commonMain/kotlin/exchange.dydx.abacus/calculator/SubAccountTransformer.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/calculator/TradeInputCalculator.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/processor/configs/FeeTierProcessor.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/processor/markets/MarketProcessor.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/processor/markets/OrderbookEntryProcessor.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/processor/markets/OrderbookProcessor.kt
#	src/commonMain/kotlin/exchange.dydx.abacus/validator/trade/TradeInputDataValidator.kt

+++ src/commonMain/kotlin/exchange.dydx.abacus/state/modal/TradingStateMachine+TransferInput.kt
@@ -110,7 +110,7 @@ fun TradingStateMachine.transfer(
                 }
                 TransferInputField.usdcSize.rawValue,
                 TransferInputField.usdcFee.rawValue -> {
-                    transfer.safeSet(typeText, parser.asDecimal(data))
+                    transfer.safeSet(typeText, parser.asDouble(data))
                     transfer.safeSet("route", null)
                     transfer.safeSet("requestPayload", null)
                     changes = StateChanges(
@@ -120,7 +120,7 @@ fun TradingStateMachine.transfer(
                     )
                 }
                 TransferInputField.size.rawValue -> {
-                    transfer.safeSet("size.size", parser.asDecimal(data))
+                    transfer.safeSet("size.size", parser.asDouble(data))
                     transfer.safeSet("route", null)
                     transfer.safeSet("requestPayload", null)
                     if (transfer["type"] == "DEPOSIT") {
